<Feat> service 계층 추가 및 match relation schema type 변경으로 인한 v2.sql append

+++ src/main/java/org/example/cowmatchingbe/domain/Stats.java
@@ -6,6 +6,6 @@ import lombok.*;
 @Entity @Table(name="stats")
 @Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
 public class Stats {
-    @Id private Integer id;
+    @Id private Long id;
     @Column(name="match_count", nullable=false) private Integer matchCount;
 }
